Loosen the HCubature element type bound for SVectors to allow Dual numbers
Interestingly, all of the other definitions were `<:Real`, so the current behavior doesn't error right away but instead it falls into a StackOverflow trying to keep changing types until it can finally call this dispatch, which it's never able to do because it has a tighter signature. Loosening this should fix this SO and allow for Dual numbers to be used with SVector inputs.